ventanilla_unica_Web_servicec_scraping/requirements.txt
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
inflect==7.5.0
iniconfig==2.1.0
isodate==0.7.2
lxml==5.3.2
more-itertools==10.6.0
packaging==24.2
platformdirs==4.3.7
pluggy==1.5.0
pyasynchat==1.0.4
pyasyncore==1.0.4
pyftpdlib==2.0.1
pytest==8.3.5
pytest-asyncio==0.26.0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
sniffio==1.3.1
soupsieve==2.6
sqlacodegen==3.0.0
typeguard==4.4.2
typing_extensions==4.12.2
urllib3==2.3.0
